Refactor translation resolvers and add new_jobs_resolver_label
Moved shared formatter imports to translations_formats module for consistency across translation resolvers. Added new_jobs_resolver_label to main_resolve, lab, and films_bot to improve job label resolution. Cleaned up __init__.py in new_resolvers and removed redundant new_jobs_resolver_label call from translations_resolvers.
