+++ ArWikiCats/main_processers/main_resolve.py
@@ -19,6 +19,7 @@ from ..make_bots.co_bots import filter_en
 from ..make_bots.format_bots import change_cat
 from ..ma_bots import ye_ts_bot
 from ..make_bots.matables_bots.bot import cash_2022
+from ..new_resolvers.new_jobs_resolver import new_jobs_resolver_label
 from ..new_resolvers.translations_resolvers import resolved_translations_resolvers
 from ..new_resolvers.translations_resolvers_v3i import resolved_translations_resolvers_v3i
 from ..new_resolvers.translations_resolvers_v2 import resolved_translations_resolvers_v2
@@ -78,6 +79,7 @@ def resolve_label(category: str, fix_label: bool=True) -> CategoryResult:
             resolved_translations_resolvers_v3i(changed_cat) or
             resolved_translations_resolvers_v2(changed_cat) or
             resolved_translations_resolvers(changed_cat) or
+            new_jobs_resolver_label(changed_cat) or
             ""
         )
