+++ ArWikiCats/make_bots/media_bots/films_bot.py
@@ -13,6 +13,7 @@ from ..matables_bots.bot import add_to_Films_O_TT, add_to_new_players
 from ..o_bots import fax
 from ...new_resolvers.translations_resolvers_v2.ministers_resolver import resolve_secretaries_labels
 from ...new_resolvers.translations_resolvers import resolved_translations_resolvers
+from ...new_resolvers.new_jobs_resolver import new_jobs_resolver_label
 # from ...translations_resolvers_v3i import resolved_translations_resolvers_v3i
 from ...new_resolvers.translations_resolvers_v2 import resolved_translations_resolvers_v2
 from .film_keys_bot import resolve_films
@@ -86,6 +87,11 @@ def te_films(category: str) -> str:
     resolved_label = resolved_translations_resolvers(normalized_category)
     if resolved_label:
         logger.info(f'>>>> (te_films) resolved_sports_formats_labels, {normalized_category=}, {resolved_label=}')
+        return resolved_label
+
+    resolved_label = new_jobs_resolver_label(normalized_category)
+    if resolved_label:
+        logger.info(f'>>>> (te_films) new_jobs_resolver_label, {normalized_category=}, {resolved_label=}')
         return resolved_label
 
     # most likely due to a circular import
