--- ArWikiCats/new_resolvers/__init__.py
@@ -1,9 +1,0 @@
-
-
-from . import translations_resolvers, translations_resolvers_v2, translations_resolvers_v3i
-
-__all__ = [
-    "translations_resolvers",
-    "translations_resolvers_v2",
-    "translations_resolvers_v3i",
-]
