+++ ArWikiCats/new_resolvers/translations_resolvers/match_labs.py
@@ -5,7 +5,7 @@ import functools
 from typing import Dict
 
 from ...helps import len_print
-from .. import FormatData
+from ..translations_formats import FormatData
 from ...translations.jobs.jobs_players_list import FOOTBALL_KEYS_PLAYERS
 from ...translations.sports.Sport_key import SPORTS_KEYS_FOR_JOBS
 from ...translations.sports.sports_lists import AFTER_KEYS
