+++ ArWikiCats/new_resolvers/translations_resolvers/us_states.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import functools
 
 from ...helps import logger
-from .. import FormatData
+from ..translations_formats import FormatData
 from ...translations import US_STATES
 
 _STATE_SUFFIX_TEMPLATES_BASE = {
