+++ ArWikiCats/new_resolvers/translations_resolvers_v2/countries_names_sport_multi_v2.py
@@ -4,7 +4,7 @@
 """
 import functools
 from ...helps import logger
-from .. import format_multi_data_v2, MultiDataFormatterBaseV2
+from ..translations_formats import format_multi_data_v2, MultiDataFormatterBaseV2
 from ...translations.nats.Nationality import all_country_with_nat_ar
 from ...translations.sports.Sport_key import SPORT_KEY_RECORDS
 from ..translations_resolvers.countries_names import formatted_data_en_ar_only
