+++ ArWikiCats/new_resolvers/translations_resolvers_v2/ministers_resolver.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import functools
 from ...helps import logger
-from .. import format_multi_data_v2, MultiDataFormatterBaseV2
+from ..translations_formats import format_multi_data_v2, MultiDataFormatterBaseV2
 from ...translations import all_country_with_nat_ar, ministers_keys
 from .nats_as_country_names import nats_keys_as_country_names
