+++ ArWikiCats/new_resolvers/translations_resolvers_v2/nats_v2.py
@@ -6,7 +6,7 @@ import functools
 
 from ...helps import logger
 from .nats_as_country_names import nats_keys_as_country_names, nats_keys_as_country_names_bad_keys
-from .. import FormatDataV2
+from ..translations_formats import FormatDataV2
 from ...translations import all_country_with_nat_ar
 from .data import country_names_and_nats_data
