+++ ArWikiCats/new_resolvers/translations_resolvers_v3i/resolve_v3i.py
@@ -20,7 +20,7 @@ from ...translations import get_from_new_p17_final, get_from_pf_keys2
 from ...make_bots.matables_bots.table1_bot import get_KAKO
 # from ..make_bots.ma_bots.country2_lab import get_lab_for_country2
 from ...helps import logger
-from .. import FormatDataFrom, MultiDataFormatterYearAndFrom
+from ..translations_formats import FormatDataFrom, MultiDataFormatterYearAndFrom
 from ..new_jobs_resolver import new_jobs_resolver_label
 
 from ...time_resolvers.time_to_arabic import convert_time_to_arabic, match_time_en_first
